requirements.txt
requests==2.24.0
pymongo==3.11.0
dnspython==2.0.0
discord.py==1.4.1
discord==1.0.1
Pillow==8.0.0
python-dotenv==0.14.0
